llmpoker-assistant/requirements.txt
# LLMPoker Assistant - Python Dependencies

# Core
python>=3.10

# Vision & ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.5.0
accelerate>=0.20.0

# FastVLM (from GitHub fork)
git+https://github.com/MrMoshkovitz/ml-fastvlm.git

# Screen capture
mss>=9.0.0
pyautogui>=0.9.54

# Web server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0.0
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0

# LLM clients
openai>=1.0.0
google-generativeai>=0.3.0

# Poker utilities
treys>=0.1.8

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
numpy>=1.24.0
requests>=2.31.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
